Refactor PDDL editor state into reusable hooks
